requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.0.8
idna==2.7
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0
django-tastypie